moving methods out of controller
