+++ apps/crn-server/test/mocks/user.controller.mock.ts
@@ -8,5 +8,4 @@ export const userControllerMock = {
   update: jest.fn(),
   updateAvatar: jest.fn(),
   syncOrcidProfile: jest.fn(),
-  createActiveCampaignContact: jest.fn(),
 } as unknown as jest.Mocked<UserController>;
