+++ apps/gp2-server/test/mocks/user.controller.mock.ts
@@ -9,8 +9,6 @@ export const userControllerMock = {
   update: jest.fn(),
   updateAvatar: jest.fn(),
   syncOrcidProfile: jest.fn(),
-  createActiveCampaignContact: jest.fn(),
-  updateActiveCampaignContact: jest.fn(),
 } satisfies jest.Mocked<
   PublicInterface<UserController>
 > as unknown as jest.Mocked<UserController>;
